requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
audioop-lts==0.2.2
cffi==1.17.1
discord.py==2.6.2
ffmpeg-python==0.2.0
frozenlist==1.7.0
future==1.0.0
idna==3.10
multidict==6.6.4
propcache==0.3.2
pycparser==2.22
PyNaCl==1.5.0
yarl==1.20.1
yt-dlp==2025.8.27
